Add ca_revoke_cert API
Added a test for the API as well to the unit test. The ca_revoke_cert
API adds an entry to the index.txt file with the revoked cert.

Thinking more on this, we should have file integrity protection for the
index.txt file. Maybe use the CA private key to sign the file and
validate it before reading and trusting it.

Can write it's contents to a plist with <data> and <signature> entries
to be validated.

Also there should be logic to not use any existing index.txt file if we
generate a new CA private key?

Maybe some challenge for the daemon to be able to prove to itself that
it has access to the private key still before trusting the index file as
well.

These are all hardening features down the line, lets get MVP working
first :)