Add ca_get_crl API
This was a bit complicated. First added a way to build the CRL from the
index.txt. Then output this to a file and as bytes to the API.

We have to hack the signature on the tbs in the X509 again because OSSL
doesn't like us calling into keychain to do the signing.